Overhaul slide controller software
Now almost everything is modularized like a proper Perl project, and is
a lot easier to follow.

The new slides will wind up using YAML to configure, and while I haven't
written an official document describing the parameters you can supply,
hopefully they're pretty self-explanatory.

Additionally, because the slides are now in single files and a lot more
internally idempotent, it's become possible to actually jump between
slides. Just press 'g', a slide number, and enter, and off you go!

Additionally cool: if you don't remember the slide number you want to go
to, you can also press 'l', and this will bring up a list of all of the
slides, with their titles.

I haven't quite made this into a way that's easy to debug when not run
in exactly this setup, but maybe that'll be on the way soon!