Fix another bug in slide jumping
It turns out that VIM::Eval() actually returns two elements: the first
is a 1 or 0 indicating whether or not the evaluated expression was
actually valid at all; this is probably handy in many cases, though I
kinda wish it were the second element instead of the first, but
whatever.

While I'm sure this return parameter is useful in many cases, it
definitely isn't for ours, so I simply opted to discard it by grabbing
the second element of the array. I could have fixed this in the
parameter list for Slide::go_to_slide() as well, but I'd prefer that
function not accept extra arguments that it always intends to discard.