Add a new slide: introducing modes
I'll have some followup slides that discuss what exactly each of these
modes mean, but I'll want to add a "file" type slide before I get into
that, because it'll be easiest to show these modes with text, instead of
on a slide of bullets. That, and I don't want the talk to be
monotonously expounding slide bullets.