Re-index slides starting at 1 instead of 0
This is how the slides are named on the filesystem, and it was starting
to get a little confusing going between the two. I could have renamed
the files, but it was easier to do this, and I think it tends to look
better anyway: only computers naturally think in 0-indexing.