Add Ctrl-R presenter control for restarting
This completely rewinds the presentation and starts over with the very
first slide. It does cause an error in the presenter display, but this
should almost never happen in real-world use anyway; it's just meant to
allow me to more easily develop new functions.

Of course, that being said, the perl functions probably won't reload
without even more work, but again, this is just an optional way to do
things.