Load slides by name only, add name to Slide object
Now the Slide object handles the construction of the full path to the
YAML file containing its specification, so that it can more easily save
the filename as a "name" parameter in its own definition. This winds up
being pretty handy for a number of reasons which will become apparent
soon.