Remove "What's Perl", replace with title slide
I both wanted to include a title slide, and I was liking the Perl joke
slide less and less, not least of which because I am actually trying to
compress time *down* now, not scale it *up* as in the first time I gave
this talk. So, rather than have a fumbling joke that's invalid and also
takes up time I don't have, I'm just replacing it with a title slide.

And, since it just happens to fall where I needed that title slide to
go, anyhow, it all works out pretty nice and neat, and I don't have to
renumber anything! I love it when a plan comes together.